pyproject.toml
[build-system]
requires = ["setuptools>=59.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ttipsu"
version = "0.0.1"
description = "Adapter for TTi PSU devices, models MX180TP & MX100QP"
readme = "README.md"
authors = [{name = "Evelyn Johns", email = "evelyn.johns@stfc.ac.uk"}]
requires-python = ">=3.9"
dependencies = ["odin-control"]

[tool.ruff]
line-length = 100
indent-width = 4
lint.select = [
    "D",    # docstrings
    "D401", # docstrings require imperative mood
    "C4",   # flake8-comprehensions - https://beta.ruff.rs/docs/rules/#flake8-comprehensions-c4
    "E",    # pycodestyle errors - https://beta.ruff.rs/docs/rules/#error-e
    "F",    # pyflakes rules - https://beta.ruff.rs/docs/rules/#pyflakes-f
    "F401",
    "W",    # pycodestyle warnings - https://beta.ruff.rs/docs/rules/#warning-w
    "I001", # isort
]

[tool.ruff.lint.pydocstyle]
convention = "google"